// src/app/models/IGitReadMe.ts
import { Link } from './link';

export interface IGitReadMe {
    name: string;
    path: string;
    sha: string;
    size: number;
    url: string;
    html_url: string;
    git_url: string;
    download_url: string;
    type: string;
    content: string;
    encoding: string;
    links: Link[];
}